refactor(v2): use react hooks on root app (#1095)
* refactor(v2): use functional component + react hooks for App

* docusaurus/link

* nits

* review
