+++ v2/lib/docusaurus/Head.js
