+++ v2/lib/theme/Pages/index.js
@@ -6,7 +6,7 @@
  */
 
 import React, {useContext} from 'react';
-import Head from '@docusaurus/head';
+import Head from '@docusaurus/Head';
 import Layout from '@theme/Layout'; // eslint-disable-line
 
 import DocusaurusContext from '@docusaurus/context';
