+++ v2/lib/theme/Sidebar/SidebarLink.js
@@ -6,18 +6,12 @@
  */
 
 import React from 'react';
-import {NavLink} from 'react-router-dom';
+import Link from '@docusaurus/Link';
 import classnames from 'classnames';
 
 import styles from './styles.module.css';
 
 export default function SidebarLink({href, label}) {
-  const isExternal = /^(https?:|\/\/)/.test(href);
-  const Link = isExternal
-    ? // eslint-disable-next-line jsx-a11y/anchor-has-content
-      ({to, activeClassName, ...linkProps}) => <a {...linkProps} href={to} />
-    : NavLink;
-
   return (
     <li className={styles.sidebarListItem}>
       <Link
