+++ v2/test/__fixtures__/transversioned-site/pages/index.js
@@ -6,7 +6,7 @@
  */
 
 import React from 'react';
-import Head from '@docusaurus/head';
+import Head from '@docusaurus/Head';
 
 export default class Home extends React.Component {
   render() {
